# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cinescore - Movie Rating Aggregator API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MovieSummary {
  tmdbId: number;
  title: string;
  /** @nullable */
  year: string | null;
  /** @nullable */
  overview: string | null;
  /** @nullable */
  posterPath: string | null;
  /** @nullable */
  backdropPath: string | null;
  genres: string[];
  /** @nullable */
  voteAverage: number | null;
  /** @nullable */
  popularity: number | null;
}

export interface RatingSource {
  source: string;
  /** @nullable */
  value: string | null;
  /** @nullable */
  normalizedScore: number | null;
}

export interface MovieDetail {
  tmdbId: number;
  /** @nullable */
  imdbId: string | null;
  title: string;
  /** @nullable */
  year: string | null;
  /** @nullable */
  overview: string | null;
  /** @nullable */
  posterPath: string | null;
  /** @nullable */
  backdropPath: string | null;
  genres: string[];
  /** @nullable */
  runtime: number | null;
  /** @nullable */
  director: string | null;
  cast: string[];
  /** @nullable */
  voteAverage: number | null;
  /** @nullable */
  popularity: number | null;
  ratings: RatingSource[];
  /** @nullable */
  movieMetricScore: number | null;
  /** @nullable */
  tagline: string | null;
  /** @nullable */
  status: string | null;
  /** @nullable */
  budget: number | null;
  /** @nullable */
  revenue: number | null;
}

export interface MovieSearchResult {
  results: MovieSummary[];
  totalResults: number;
  page: number;
  totalPages: number;
}

export interface MovieReview {
  review: string;
  generatedAt: string;
}

export interface RecommendationsResult {
  results: MovieSummary[];
}

export type SearchMoviesParams = {
  /**
   * Search query
   */
  q: string;
  page?: number;
};