# lib/api-zod/src/generated/types/movieReview.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cinescore - Movie Rating Aggregator API
 * OpenAPI spec version: 0.1.0
 */

export interface MovieReview {
  review: string;
  generatedAt: string;
}